+++ src/scicloj/metamorph/ml/classification.clj
@@ -104,4 +104,4 @@
       (ds/new-dataset [(ds/new-column target-column-name dummy-labels {:column-type :prediction
                                                                        :categorical-map (get target-categorical-maps target-column-name)})])))
      
-  {})
+  {:glance-fn (fn [_] (ds/->dataset {:npar 0}))})